\name{RFPerm-package}
\alias{RFPerm-package}
\alias{RFPerm}
\docType{package}
\title{
\packageTitle{RFPerm}
}
\description{
\packageDescription{RFPerm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RFPerm}
\packageIndices{RFPerm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RFPerm}

Maintainer: \packageMaintainer{RFPerm}
}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
