\name{RFTest-package}
\alias{RFTest-package}
\alias{RFTest}
\docType{package}
\title{
\packageTitle{RFTest}
}
\description{
\packageDescription{RFTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RFTest}
\packageIndices{RFTest}
Most of the functionality comes in the \code{MSE_Test} file, which will conduct an F-test for a certain subset of variables. To run marginal F-tests for each variable, \code{f_holdoutRF} (efficient) or \code{permtestImp} (brute force) should be used.
}
\author{
\packageAuthor{RFTest}

Maintainer: \packageMaintainer{RFTest}
}
\references{

}
\keyword{ package }

