\name{RFPerm-package}
\alias{RFPerm-package}
\alias{RFPerm}
\docType{package}
\title{
\packageTitle{RFPerm}
}
\description{
\packageDescription{RFPerm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RFPerm}
\packageIndices{RFPerm}
Most of the functionality comes in the \code{MSE_Test} file, which will conduct an F-test for a certain subset of variables. To run marginal F-tests for each variable, \code{f_holdoutRF} (efficient) or \code{permtestImp} (brute force) should be used.
}
\author{
\packageAuthor{RFPerm}

Maintainer: \packageMaintainer{RFPerm}
}
\references{

}
\keyword{ package }

